🔧 [Update]: Add SupportsShouldProcess to multiple functions for improved safety and user confirmation during execution.
🔧 [Patch]: Rename functions for module consistency #12

+++ src/functions/private/Remove-WWAppx.ps1
@@ -23,7 +23,7 @@
     Remove-WWAppx -array $exappxs
 #>
 function Remove-WWAppx {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param(
         [Parameter(Mandatory = $true)]
         [array]$array
@@ -33,12 +33,14 @@ function Remove-WWAppx {
         $exappxs = $array
         Write-WimWitchLog -data 'Starting AppX removal' -class Information
         foreach ($exappx in $exappxs) {
-            try {
-                Remove-AppxProvisionedPackage -Path $WPFMISMountTextBox.Text -PackageName $exappx -ErrorAction Stop | Out-Null
-                Write-WimWitchLog -data "Removing $exappx" -Class Information
-            } catch {
-                Write-WimWitchLog -Data "Failed to remove $exappx" -Class Error
-                Write-WimWitchLog -Data $_.Exception.Message -Class Error
+            if ($PSCmdlet.ShouldProcess($exappx, "Remove AppX Package")) {
+                try {
+                    Remove-AppxProvisionedPackage -Path $WPFMISMountTextBox.Text -PackageName $exappx -ErrorAction Stop | Out-Null
+                    Write-WimWitchLog -data "Removing $exappx" -Class Information
+                } catch {
+                    Write-WimWitchLog -Data "Failed to remove $exappx" -Class Error
+                    Write-WimWitchLog -Data $_.Exception.Message -Class Error
+                }
             }
         }
         return
