+++ src/functions/private/Update-WWWindowsRe.ps1
@@ -23,39 +23,36 @@
     Update-WWWindowsRe
 #>
 function Update-WWWindowsRe {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param(
 
     )
 
     process {
         try {
-
-    #create mount point in staging
-    #copy winre from mounted offline image
-    #change attribute of winre.wim
-    #mount staged winre.wim
-    #update, dismount
-    #copy wim back to mounted offline image
-
             Write-WimWitchLog -Data 'Starting WinRE WIM update process...' -Class Information
 
             $mountPath = $WPFMISMountTextBox.Text
             $winREPath = Join-Path -Path $mountPath -ChildPath 'Windows\System32\Recovery\WinRE.wim'
 
             if (Test-Path -Path $winREPath) {
-                # Mount WinRE
                 $winREMount = Join-Path -Path $mountPath -ChildPath 'WinREMount'
-                New-Item -Path $winREMount -ItemType Directory -Force | Out-Null
 
-                Mount-WindowsImage -ImagePath $winREPath -Index 1 -Path $winREMount
+                if ($PSCmdlet.ShouldProcess($winREMount, "Create mount directory")) {
+                    New-Item -Path $winREMount -ItemType Directory -Force | Out-Null
+                }
+
+                if ($PSCmdlet.ShouldProcess($winREPath, "Mount WinRE.wim")) {
+                    Mount-WindowsImage -ImagePath $winREPath -Index 1 -Path $winREMount
+                }
 
                 # Perform updates here
                 # Add specific WinRE update logic
 
-                # Dismount WinRE
-                Dismount-WindowsImage -Path $winREMount -Save
-                Remove-Item -Path $winREMount -Force
+                if ($PSCmdlet.ShouldProcess($winREMount, "Dismount WinRE.wim")) {
+                    Dismount-WindowsImage -Path $winREMount -Save
+                    Remove-Item -Path $winREMount -Force
+                }
 
                 Write-WimWitchLog -Data 'WinRE WIM updated successfully' -Class Information
             }
